Web-demo/requirements.txt
transformers
sentence-transformers
faiss-cpu
pandas
pyarrow
torch
pillow
google-generativeai
gradio